hotels: split ComparablePrice invariant tests into currency_normalize_test.go
The three ensureComparableInTargetCurrency regression tests pushed
search_extra_test.go to 814 lines, past the 800-line CI hygiene ceiling.
They belong beside the code they exercise anyway: move them into a new
currency_normalize_test.go (the test file for currency_normalize.go, where
ensureComparableInTargetCurrency lives). No test logic changes — a pure
relocation. search_extra_test.go drops to 762 lines; the new file is 59.
